feat: introduce abstract class url3d
