+++ docarray/typing/url/__init__.py
@@ -1,7 +1,7 @@
 from docarray.typing.url.any_url import AnyUrl
 from docarray.typing.url.image_url import ImageUrl
-from docarray.typing.url.mesh_url import Mesh3DUrl
-from docarray.typing.url.point_cloud_url import PointCloud3DUrl
 from docarray.typing.url.text_url import TextUrl
+from docarray.typing.url.url_3d.mesh_url import Mesh3DUrl
+from docarray.typing.url.url_3d.point_cloud_url import PointCloud3DUrl
 
 __all__ = ['ImageUrl', 'AnyUrl', 'TextUrl', 'Mesh3DUrl', 'PointCloud3DUrl']
