+++ docarray/typing/url/url_3d/__init__.py
@@ -1,0 +1,4 @@
+from docarray.typing.url.url_3d.mesh_url import Mesh3DUrl
+from docarray.typing.url.url_3d.point_cloud_url import PointCloud3DUrl
+
+__all__ = ['Mesh3DUrl', 'PointCloud3DUrl']
